Widen NodeId to u64
A 32-bit process identifier is a needless ceiling on how identities can
be assigned — derived from a hash, an address, or an external registry
rather than counted from zero. Nothing depended on the width.

+++ crates/recon-core/src/node.rs
@@ -7,10 +7,10 @@ use core::fmt;
 /// Copy and totally ordered, so collections keyed by it iterate deterministically and
 /// membership sets can be enumerated in a stable order.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct NodeId(pub u32);
+pub struct NodeId(pub u64);
 
 impl NodeId {
-    pub const fn new(n: u32) -> Self {
+    pub const fn new(n: u64) -> Self {
         NodeId(n)
     }
 }
